+++ src/main/resources/application-prod.yml
@@ -36,45 +36,3 @@ logging:
     org.springframework.transaction.interceptor: info
 
 remote-addr: 133.186.134.30
-
-#    shutdown: graceful
-#
-#  eureka:
-#    instance:
-#      instance-id: coupon
-#      hostname: localhost
-#      prefer-ip-address: true
-#    client:
-#      fetch-registry: true
-#      register-with-eureka: true
-#      service-url:
-#        defaultZone: http://admin:1234@localhost:8761/eureka
-#
-#  spring:
-#    lifecycle:
-#      timeout-per-shutdown-phase: 30s
-#
-#  management:
-#    health:
-#      status:
-#        order: DOWN, UP
-#    endpoint:
-#      jolokia:
-#        enabled: true
-#      metrics:
-#        enabled: true
-#      pause:
-#        enabled: true
-#      resume:
-#        enabled: true
-#      restart:
-#        enabled: true
-#      shutdown:
-#        enabled: true
-#    endpoints:
-#      web:
-#        exposure:
-#          include: '*'
-#    info:
-#      env:
-#        enabled: true
